Create CODE_OF_CONDUCT.md
I have my doubts on this ever being useful.

But, I'm still deciding when `commit` should collect its file list: at add-time allows this to be done without adding a function to `crossplatform.h`, but requires storing a file in the `.vcs` directory and manual deduplication (to prevent double-adding filenames), while at commit-time allows the OS to track files (well, the filesystem, but the OS is the one that tells us about it), and since this is only checked once during, commit duplication cannot occur, but this incurs the need to write a shim to generate a vector of filenames in `crossplatform.h` (since Windows and *nix do this differently pre-C++17)

Since I haven't decided on which path to take yet, I'm procrastinating by working on contribution guidelines. This way, I can put off making a decision, while still technically working on the project.

Sue me.