Fix commit
... I'm retarded.

No, really, I am literally retarded, apparently.

When writing commit, I was thinking about how I needed binary files to
work properly, so one could commit a JPEG or something like that. Which is
why the files to be committed are read binary.

.... And then I managed to /not/ write the commit as a binary file,
ruining all the work I put in. Which of course breaks all sorts of things,
because the file undergoes the binary-text transformation through the save
operation.

Holy hell. I've been trying to figure this out for days, and /this/ is the
problem? Goddamnit...

*sigh*

Even though the problem with checkout (which caused the associated test
script to fail) wasn't actually in checkout, thanks to this commit
checkout and commit are both reading the same formats. Therefore, I hereby
declare...

This commit fixes #9.

Signed-off-by: Ryan Hodin <ryan951357@gmail.com>