Update date submodule
This changes the directory structure for the date header, and I hate it.

It was done in order to "make it easier to include files when using submodules", but as the diff for hero.cpp here shows, it has certainly not done so for Hero. Instead of "date/<header>.h", you now include "date/include/date/<header>.h", which in my opinion is worse in every possible way.

But, its not my project, so I'll live with it. Its not that big a deal, I'll forget about it by tomorrow, but today it bothers me.