Make the repository path ".hero/"
Long, long ago (by which I mean just under two months ago, in
b7194588f58319c528b4708b6243c0dc6a973043), back when the project was named
VersionControl (hence the relic in the paths), I decided to use the folder
".vcs" to store repository data. This, of course, like the name of the
project, was a placeholder, and it was short for ".VersionControlSystem".

Many eons (almost exactly one month) later, in
5e1abd4bac1ec8b2f48490c30ceba05b05934ff7, the name of the project was
officially formally changed to "Hero", and there was much rejoicing, as
this was considered an actual appropriate name for the project.

This, however, left a relic, for "hero" now stored its data in ".vcs", and
the dragon called Confusion reared its ugly head, however small, for those
few who paid attention were left scratching their less ugly heads.

Today, the great evils known as Laziness and Procrastination were driven
back, and the great soldier who wrote this damn thing got off his ass, and
took arms against the dragon called Confusion, and by opposing, ended it.

Okay, in other words, I switched ".vcs" to ".hero", so now repository data
is stored in a folder named after the program again. That's not as fun
though, is it?