Fix errors related to appended
You see, I forgot that the return value of c_str() would go out of scope
with the string it came from.

To fix the problem, I went down the rabbit hole, and ended up basically
making a mini-std::string that implicitly converts to c-strings.

Ugh. At least it compiles. At least it compiles.... And works.