Add the hero-repofix project
This project is a specific binary for upgrading a repository made with an
old version of hero to the current version.

I chose to make this its own binary, rather than adding it as a command to
the main `hero` binary (hero repofix) in order to provide a clear point of
segmentation where the main binary has nothing to do with actually dealing
with old formats.

I do, however, think that I'll allow `hero repofix` to invoke
`hero-repofix`.