Add indexmap class
This class will be used to store a table of the filenames in the
index against their hashes, to be used for directory support (for example,
to prevent two versions of the same file from being committed twice).

Lacking are stream operations. These will be added shortly, as they are
critical.