Define an indexmap path constant
So that, if I change my mind about the storage for the indexmap later, its
easy to change.

Also, in the near future, either before the next version or before the one
after that, Hero is going to have to go through a pretty massive code
rework, because the structure right now is pretty bad and I have a decent
amount of technical debt to work with from too long spent doing small
'touch once and never again' projects. What's going to happen is that I'm
going to split everything up into a set of classes 'Repository', 'Commit',
'Index', 'CommitFile' and so on. These classes will naturally be virtual,
probably with the base class being abstract, so that they can be
implemented as necessary (that is, we could have a 'DifferenceCommitFile'
and a 'BaseCommitFile' to implement diffencoding, inheriting from a base
that recognizes which to generate.

That means that, eventually, 'Repository' will do the internal work of
handling Indexmap storage, so this structure resembles its eventual
implementation.