Implement a template loader for indexmaps
Thus providing an RAII-compliant interface for containing indexmaps which
should always stay synchronized with their on-disk equivalents. This is
not crash-safe: If the program is interrupted, the last changes will not
be written. Another interface with more integration with Indexmap would
have to be written to allow for more synchronization if this turns out to
be absolutely necessary.