Finish implementing indexmap support for commit
There wasn't actually as much as I thought there would be.

There's actually quite a bit of room for improvement with how these files
are iterated over. A tremendous amount. I was going to leave it for later,
but now that I'm writing this message I can't stand to leave it in, so
I'll be redoing this in a moment.

First, there is one thing I would like to document: The file hash that
gets stored in the commit, while it should always be the same as the one
stored in the indexmap, is recalculated for the commit. This is to make
double-sure that the commit has correct data, as the entire point of it is
to preserve data for the future.
I'll be adding a warning if there's a mismatch, also in a moment.