Iterate over Commitmap directly
Instead of using an in-between list of filenames like we used to, directly use the data loaded from the indexmap.

While loading and parsing the indexmap is probably slower than listing files in the index (through filesInDirectory), doing it this way is probably faster than doing both (since we need the related filenames from the indexmap anyway), and this removes the need to keep a list of filenames in memory outside of Commitmap, and this removes a lot of map lookups.

This does mean that the biggest shames are commitLast and commitFiles, both of which (by virtue of calling add immediately followed by commit) unnecessarily write the indexmap out of memory then read it back from disk.

This, however, would be a major item to restructure. So it won't be changing anytime soon.