Refactor add()
A separate function performs the recursion, and all of these functions
share an Indexmap. This prevents the maps from interfering with each
other, and prevents a flood of "All files added to index." on recursive
adds

By making directory support work, and by finishing the seconds
quality-of-life feature I wanted to add, this commit fixes #10.

Signed-off-by: Ryan Hodin <ryan951357@gmail.com>